requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-binance>=1.0.0

# Cryptography for API key encryption
cryptography>=3.4.0

# GUI (optional)
matplotlib>=3.5.0
mplfinance>=0.12.7a0

# Logging
colorlog>=6.0.0

# Technical analysis
ta>=0.10.0

# Configuration
python-dotenv>=0.19.0

# Development and testing (optional)
pytest>=6.0.0
pytest-cov>=3.0.0